# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample chat service API.",
        "title": "Chat Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Wellington Ferreira",
            "url": "https://github.com/wrferreira1003",
            "email": "wellingtonferreira1003@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/chat/messages/{userID1}/{userID2}": {
            "get": {
                "description": "Obtem as mensagens de uma conversa",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Obtem as mensagens de uma conversa",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do primeiro usuário",
                        "name": "userID1",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID do segundo usuário",
                        "name": "userID2",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de mensagens",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Message"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/chat/send": {
            "post": {
                "description": "Envia uma mensagem para o chat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Envia uma mensagem para o chat",
                "parameters": [
                    {
                        "description": "Mensagem a ser enviada",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Message"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Mensagem enviada com sucesso",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Erro na requisição",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "conversation_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "receiver_id": {
                    "type": "string"
                },
                "sender_id": {
                    "type": "string"
                }
            }
        }
    }
}